Migrate from GitHub Pages to Cloudflare Pages
This switch is done because:
- There is no standard way to config GitHub pages with redirection/forward rules.
Only some dirty workarounds exist.
Thus, it is not suitable for SPAs.
- There is no standard way to modify cache-control policies.
The default value is just 10 minutes for all the resources.

Perfect SPA migration candidates:
- Cloudflare Pages: minimal, intuitive, performant & reliable, workers
- Vercel: feature-packed, intuitive, performant & reliable, functions
- Netlify: feature-packed, intuitive, performant & reliable, functions
- Render: minimal, full-stack, PaaS, no functions

Other SPA migration candidates:
- statichost.eu: minimal, privacy-friendly, performance and reliability concerns
- Firebase Hosting: Google ecosystem, not available from Iranian IPs
- DigitalOcean App Platform: full-stack, credit-card required
- Azure SPA: Azure Ecosystem, credit card and identification required
- AWS Amplify: AWS Ecosystem, free tier for only 12 months, credit card and identification required
- GitLap Pages: My code is on GitHub!

After thorough evaluation of the above options and deploying my website on most of them, I decided to go with Cloudflare Pages because of minimalism and performance, but you really can't go wrong with any of the top 4 options.